+++ src/main/java/pl/lukaszgrymulski/noteservice/repository/NoteRepository.java
@@ -16,10 +16,10 @@ public interface NoteRepository extends JpaRepository<NoteEntity, LocalDateTime>
     List<NoteEntity> findAllRecentNoteVersions();
 
     @Query(value = "SELECT n FROM NoteEntity n WHERE n.is_deleted=0 and n.id=:id")
-    Optional<NoteEntity> findRecentNoteVersionById(int id);
+    Optional<NoteEntity> findRecentNoteVersionById(Long id);
 
     @Query(value = "SELECT MAX(n.id) FROM NoteEntity n")
-    int findMaxId();
+    Long findMaxId();
 
-    List<NoteEntity> findAllById(int id);
+    List<NoteEntity> findAllById(Long id);
 }
